requirements.txt
streamlit
langchain-core
langchain-openai
pydantic
python-dotenv
python-docx